+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Make `as_slice()/from_slice()` const.
     * Make `ArrayLength: 'static`
     * Replace `From<&[T]>` with `TryFrom<&[T]>`
+    * `Deserialize` no longer requires `T: Default`
 
 * **`0.14.6`**
     * Add an optional `Zeroize` impl for `GenericArray` ([#126](https://github.com/fizyk20/generic-array/pull/126) and [#112](https://github.com/fizyk20/generic-array/pull/112))